chore: Move callback resolver to tx confirmation
