+++ source/pages/Transactions/Transaction.tsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 
 import { Layout } from 'components/index';
 import { useQueryData } from 'hooks/index';
-import { getController } from 'utils/browser';
 
 import Fee from './Fee';
 import TransactionConfirmation from './TransactionConfirmation';
@@ -29,37 +28,6 @@ const titleResolver = (txType: string) => {
   }
 };
 
-const callbackResolver = (txType: string) => {
-  let callbackName;
-
-  switch (txType) {
-    case 'CreateToken':
-      callbackName = 'confirmTokenCreation';
-      break;
-
-    case 'CreateNFT':
-      callbackName = 'confirmNftCreation';
-      break;
-
-    case 'MintToken':
-      callbackName = 'confirmTokenMint';
-      break;
-
-    case 'MintNFT':
-      callbackName = 'confirmMintNFT';
-      break;
-
-    case 'UpdateToken':
-      callbackName = 'confirmUpdateToken';
-      break;
-
-    default:
-      throw new Error('Unknown transaction type');
-  }
-
-  return getController().wallet.account.sys.tx[callbackName];
-};
-
 interface ITransaction {
   type: string;
 }
@@ -72,7 +40,6 @@ const Transaction: React.FC<ITransaction> = ({ type }) => {
   const [fee, setFee] = useState<number>();
 
   const title = titleResolver(type);
-  const callback = callbackResolver(type);
 
   if (!fee) return <Fee title={title} onFinish={setFee} />;
 
@@ -83,7 +50,6 @@ const Transaction: React.FC<ITransaction> = ({ type }) => {
         title={title}
         type={type}
         transaction={{ ...transaction, fee }}
-        callback={callback}
       />
     </Layout>
   );
